university_portal/requirements.txt
ABN==1.0.0
absl-py==2.0.0
asgiref==3.8.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.8
Django==5.1
django-jalali==7.1.0
django-resized==1.0.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.4.2
filelock==3.15.4
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
jalali_core==1.0.0
jdatetime==5.0.0
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlip==0.2
ml-dtypes==0.2.0
mypy==1.7.1
mypy-extensions==1.0.0
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
platformdirs==4.2.2
protobuf==4.23.4
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.9.0
pymongo==4.6.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.1.0
virtualenv==20.26.3
Werkzeug==3.0.1
wrapt==1.14.1
